a2-eventos/backend/camera-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
numpy>=1.26.0
opencv-python>=4.9.0.80
pydantic>=2.6.0
loguru>=0.7.2
python-dotenv>=1.0.1
Pillow>=10.2.0
httpx>=0.26.0
psycopg2-binary>=2.9.9
redis>=5.0.1
insightface>=0.7.3
ultralytics>=8.1.0
easyocr>=1.7.1
python-multipart>=0.0.9
websockets>=12.0
python-jose[cryptography]>=3.3.0